requirements.txt
fastapi>=0.68.0
gradio>=3.50.2
google-generativeai>=0.3.0
transformers>=4.30.0
langchain>=0.1.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.23
pandas>=1.3.0
pytest>=6.2.5
uvicorn>=0.15.0
black>=22.3.0
pre-commit>=2.17.0
pydantic-settings>=2.0.0
